UPDATE: new SQL schema (again)

+++ php/removeData.php
@@ -10,12 +10,12 @@ $s = $db->prepare($q);
 $s->execute([ 'id' => $id ]);
 $r = $s->fetchAll();
 
-if ($r[0]['sDate'] == $date and $r[0]['fDate'] == $date and $r[0]['real_evt_id'] == NULL and $r[0]['real_sDate'] == NULL) {
+if ($r[0]['sDate'] == $date and $r[0]['fDate'] == $date and $r[0]['real_evt_id'] == NULL and $r[0]['real_sDate'] == NULL and $r[0]['real_fDate'] == NULL) {
     // delete real_evt
     // $q1 = "DELETE FROM `event` WHERE `event`.`id` = :id;";
     // $s1 = $db->prepare($q1);
     // $s1->execute([ 'id' => $id ]);
-} else if ($r[0]['sDate'] == $date and $r[0]['fDate'] != $date and $r[0]['real_evt_id'] == NULL and $r[0]['real_sDate'] == NULL) {
+} else if ($r[0]['sDate'] == $date and $r[0]['fDate'] != $date and $r[0]['real_evt_id'] == NULL and $r[0]['real_sDate'] == NULL and $r[0]['real_fDate'] == NULL) {
     // delete end_evt
     $q1 = "DELETE FROM `event` WHERE `event`.`real_evt_id` = :id;";
     $s1 = $db->prepare($q1);
@@ -24,14 +24,14 @@ if ($r[0]['sDate'] == $date and $r[0]['fDate'] == $date and $r[0]['real_evt_id']
     // $q2 = "DELETE FROM `event` WHERE `event`.`id` = :id;";
     // $s2 = $db->prepare($q2);
     // $s2->execute([ 'id' => $id ]);
-} else if ($r[0]['sDate'] == $date and $r[0]['fDate'] == $date and $r[0]['real_evt_id'] != NULL and $r[0]['real_sDate'] != NULL) {
+} else if ($r[0]['sDate'] == $date and $r[0]['fDate'] == $date and $r[0]['real_evt_id'] != NULL and $r[0]['real_sDate'] != NULL and $r[0]['real_fDate'] != NULL) {
     $sD = DateTime::createFromFormat('Y-m-d', $r[0]['sDate']);
     $r_sD = DateTime::createFromFormat('Y-m-d', $r[0]['real_sDate']);
     $new_fD = DateTime::createFromFormat('Y-m-d', $r[0]['fDate']);
     $diff = $sD->diff($r_sD)->d;
     if ($diff === 1) {
         // check n end_evt
-        $q0 = "SELECT COUNT(`id`) FROM `event` WHERE `event`.`real_evt_id` = :r_id";
+        $q0 = "SELECT COUNT(`id`) FROM `event` WHERE `event`.`real_evt_id` = :r_id;";
         $s0 = $db->prepare($q0);
         $s0->execute([
             'r_id' => $r[0]['real_evt_id'],
@@ -54,7 +54,7 @@ if ($r[0]['sDate'] == $date and $r[0]['fDate'] == $date and $r[0]['real_evt_id']
                 'r_id' => $r[0]['real_evt_id'],
             ]);
             // update other end-evt
-            $q2 = "UPDATE `event` SET `real_sDate` = :new_date WHERE `event`.`real_evt_id` = :r_id and `event`.`real_sDate` IS NOT NULL";
+            $q2 = "UPDATE `event` SET `real_sDate` = :new_date WHERE `event`.`real_evt_id` = :r_id and `event`.`real_sDate` IS NOT NULL;";
             $s2 = $db->prepare($q2);
             $s2->execute([
                 'new_date' => $r[0]['sDate'],
@@ -73,11 +73,18 @@ if ($r[0]['sDate'] == $date and $r[0]['fDate'] == $date and $r[0]['real_evt_id']
             'r_id' => $r[0]['real_evt_id'],
         ]);
         // delete end-evts successivi
-        $q2 = "DELETE FROM `event` WHERE `event`.`real_evt_id` = :r_id and `event`.`sDate` > :limit_date";
+        $q2 = "DELETE FROM `event` WHERE `event`.`real_evt_id` = :r_id and `event`.`sDate` > :limit_date OR `event`.`fDate` > :limit_date;";
         $s2 = $db->prepare($q2);
         $s2->execute([
             'r_id' => $r[0]['real_evt_id'],
             'limit_date' => $str_fDate,
+        ]);
+        // update end-evts rimasti
+        $q3 = "UPDATE `event` SET `real_fDate` = :new_date WHERE `event`.`real_evt_id` = :r_id;";
+        $s3 = $db->prepare($q3);
+        $s3->execute([
+            'new_date' => $str_fDate,
+            'r_id' => $r[0]['real_evt_id'],
         ]);
         // delete end-evt selected outside of ELSE-IF
     }
